requirements.txt
efficientnet
numpy
opencv-contrib-python-headless
streamlit
<<<<<<< HEAD
tensorflow-cpu==2.1.0
=======
tensorflow-cpu==2.2.0
>>>>>>> parent of b14a3a3... Update requirements.txt



